+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-30T11:22:44.854Z",
+    "2026-01-30T11:55:17.562Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {