+++ packages/orq-rc/RELEASES.md
@@ -6279,3 +6279,13 @@ Based on:
 - [typescript v4.3.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.12 - packages/orq-rc
+
+## 2026-01-30 11:58:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.13 - packages/orq-rc